+++ sshd-core/src/test/resources/org/apache/sshd/client/auth/pubkey/entrypoint.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+chown -R bob /home/bob
+chmod 0600 /home/bob/.ssh/*
+/usr/sbin/sshd -D -ddd